requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
discord==2.3.2
discord.py==2.4.0
docutils==0.21.2
et-xmlfile==1.1.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
lockfile==0.12.2
multidict==6.0.5
numpy==2.1.0
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.2
pid==3.0.4
psutil==6.0.0
pycparser==2.22
pyperclip==1.9.0
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
selenium==4.23.1
service==0.6.0
setproctitle==1.3.3
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zope.interface==7.0.1